adjust_task/requirements.txt
Django==3.2.1
mysqlclient==1.4.6
djangorestframework==3.12.2
django-filter==2.3.0
ddt==1.4.2
drf-yasg==1.20.0